requirements.txt
fastapi
uvicorn
qdrant-client
httpx
pypdf
python-dotenv
python-multipart
together 
